pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.9", "pybind11>=2.12"]
build-backend = "scikit_build_core.build"

[project]
name = "pylitehtml"
version = "0.1.0"
description = "HTML+CSS to image renderer based on litehtml"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = ["jinja2>=3.0"]
classifiers = [
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
  "Typing :: Typed",
]

[project.optional-dependencies]
dev = ["pytest", "pillow", "numpy", "pytest-asyncio"]

[tool.scikit-build]
cmake.build-type = "Release"
wheel.packages = ["src/python/pylitehtml"]
cmake.verbose = true

[tool.scikit-build.cmake.define]
LITEHTML_BUILD_TESTING = "OFF"

[tool.ruff]
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
